Kleine refactors, audio types/models toegevoegd, opzetje manager
