+++ Assets/Audio/SFX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0607d26eb2cbb643bac2b80c29692cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
