+++ Assets/Prefabs/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c87bad1ca42464548bb60d85965980f0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
