+++ Assets/Scripts/Enums.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97cc8d703dc577944a2ab0e4eff6a8e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
