+++ Assets/Scripts/Enums/BGMType.cs
@@ -1,0 +1,5 @@
+
+    public enum BGMType
+    {
+        PlaceholderTrack
+    }
