+++ Assets/Scripts/Enums/BGMType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9976169d64141b74bab13c57af9e2877
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
