+++ Assets/Scripts/Enums/SFXType.cs
@@ -1,0 +1,7 @@
+
+    public enum SFXType
+    {
+        PlayerTakeDamage,
+        PlayerRangedAttack,
+        PlayerMeleeAttack
+    }
