+++ Assets/Scripts/Enums/SFXType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 402e21e7f020cc94891045e8861be03b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
