+++ Assets/Scripts/Enums/UpgradeType.cs
@@ -1,0 +1,13 @@
+public enum UpgradeType
+{
+    none = 0,
+    rangedDamage = 1,
+    meleeDamage = 2,
+    rangedPierce = 3,
+    summonSpeed = 4,
+    summonLevel = 5,
+    meleeAoe = 6,
+    health = 7,
+    regen = 8,
+    chainLighting = 9
+}
