+++ Assets/Scripts/Models/BGMAudioClip.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[System.Serializable]
+public class BGMAudioClip : BaseAudioClip
+{
+    public BGMType BGMType;
+}
