+++ Assets/Scripts/Models/BGMAudioClip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 330c4eaba39a17b43806718087b5670b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
