+++ Assets/Scripts/Models/BaseAudioClip.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+
+public abstract class BaseAudioClip 
+{        
+    public AudioClip AudioClip;
+}
