+++ Assets/Scripts/Models/BaseAudioClip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af82a04a43c763944aeab45d3a61239b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
