+++ Assets/Scripts/Models/SFXAudioClip.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[System.Serializable]
+public class SFXAudioClip : BaseAudioClip
+{
+    public SFXType SfxType;
+}
