+++ Assets/Scripts/Models/SFXAudioClip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a643115d67cba9f4ca80d24882775322
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
